Move createNewStory to actions file
