+++ src/app/catalog/[story]/new/page.tsx
@@ -68,7 +68,6 @@ async function createNewChapter(formData: FormData) {
   }
 }
 
-
 async function getStoryId(slug: string) {
   const response = await sql.query(
     "SELECT story_id FROM stories WHERE slug = $1",
