+++ src/lib/actions/chapterActions.ts
@@ -1,2 +1,2 @@
-"use server"
+"use server";
 // This file will contain the create, edit, and delete actions for all chapter related routes
